This block of text is present only in Test1.txt.
This block of text is present only in Test1.txt.
This block of text is present only in Test1.txt.

This block of text is present in both files.
This block of text is present in both files.
This block of text is present in both files.

This block of text is present in both files.
This block of text is present in both files.
This block of text is present in both files.

This block of text has been modified between the files.
This block of text has been modified between the files.
This block of text has been modified between the files.

This block of text is present in both files.
This block of text is present in both files.
This block of text is present in both files.

This block of text differs only in whitespace. Whether
differences are highlighted depends on how the application
is configured.
Space at start of line in Test2.txt.
Tab at start of line in Test2.txt.
Space and tabs within the line in Test2.txt.

This block of text is present in both files.
This block of text is present in both files.
This block of text is present in both files.

This block of text differs only in line endings. In
Test1.txt, lines end with a CR LF pair. In Test2.txt,
lines end with LF only. Whether differences are highlighted
depends on how the application is configured.

This block of text is present in both files.
This block of text is present in both files.
This block of text is present in both files.

This block of text differs only in character case. Whether
differences are highlighted depends on how the application
is configured.